done first part of 3rd module
