+++ js/components/game-3.js
@@ -1,9 +1,9 @@
-import {getElementFromTemplate} from "./getElement";
-import {appPageShow} from "./appPageShow";
-import game3 from "./game-3";
+import {getElementFromTemplate} from "../data/getElement";
+import {appPageShow} from "../data/appPageShow";
+import stats from "./stats";
 import intro from "./intro";
 
-const template = getElementFromTemplate(`<div id="game-2">
+const template = getElementFromTemplate(`<div id="game-3">
   <header class="header">
     <div class="header__back">
       <button class="back">
@@ -19,18 +19,16 @@ const template = getElementFromTemplate(`<div id="game-2">
     </div>
   </header>
   <div class="game">
-    <p class="game__task">Угадай, фото или рисунок?</p>
-    <form class="game__content  game__content--wide">
+    <p class="game__task">Найдите рисунок среди изображений</p>
+    <form class="game__content  game__content--triple">
       <div class="game__option">
-        <img src="http://placehold.it/705x455" alt="Option 1" width="705" height="455">
-        <label class="game__answer  game__answer--photo">
-          <input name="question1" type="radio" value="photo">
-          <span>Фото</span>
-        </label>
-        <label class="game__answer  game__answer--wide  game__answer--paint">
-          <input name="question1" type="radio" value="paint">
-          <span>Рисунок</span>
-        </label>
+        <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
+      </div>
+      <div class="game__option  game__option--selected">
+        <img src="http://placehold.it/304x455" alt="Option 2" width="304" height="455">
+      </div>
+      <div class="game__option">
+        <img src="http://placehold.it/304x455" alt="Option 3" width="304" height="455">
       </div>
     </form>
     <div class="stats">
@@ -62,21 +60,13 @@ const template = getElementFromTemplate(`<div id="game-2">
 
 export default () => {
   const currentPage = template.cloneNode(true);
-  currentPage.addEventListener(`click`, () => {
-    const gameQuestion = currentPage.querySelectorAll(`.game__option`);
-    let isAllAnswered = true;
+  currentPage.querySelector(`.game__content`).addEventListener(`click`, (event) => {
+    const gameAnswer = currentPage.querySelectorAll(`.game__option`);
 
-    for (let i = 0; i < gameQuestion.length; i++) {
-      const gameSelect = gameQuestion[i].querySelectorAll(`label input:checked`);
-      const isAnswered = gameSelect.length;
-
-      if (!isAnswered) {
-        isAllAnswered = false;
-        break;
+    for (let i = 0; i < gameAnswer.length; i++) {
+      if (gameAnswer[i] === event.target) {
+        appPageShow(stats());
       }
-    }
-    if (isAllAnswered) {
-      appPageShow(game3());
     }
   });
   currentPage.querySelector(`.back`).addEventListener(`click`, () => {
