--- js/data/game-data.test.js
@@ -1,9 +1,0 @@
-import {assert} from 'chai';
-
-describe(`Array`, () => {
-  describe(`#indexOf()`, () => {
-    it(`should return -1 when the value is not present`, () => {
-      assert.equal(-1, [1, 2, 3].indexOf(4));
-    });
-  });
-});
