+++ js/data/getTimer.test.js
@@ -1,0 +1,13 @@
+import {assert} from "chai";
+import getTimer from "./getTimer.js";
+
+describe(`Timer:`, () => {
+
+  it(`Timer should be a number`, () => {
+    assert.throws(() => getTimer(``));
+  });
+
+  it(`Timer should be positive`, () => {
+    assert.throws(() => getTimer(`-6`));
+  });
+});
